// apps/electron/src/renderer/src/gen/api/schemas/getApiV1TodosParams.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Techoo API
 * API for the Techoo task management application with OpenAPI documentation
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiV1TodosDone } from './getApiV1TodosDone'

export type GetApiV1TodosParams = {
  /**
   * Range start (UTC Unix timestamp)
   * @nullable
   */
  from?: number | null
  /**
   * Range end (UTC Unix timestamp)
   * @nullable
   */
  to?: number | null
  /**
   * Filter by completion status
   */
  done?: GetApiV1TodosDone
}